% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fcPainter.R
\name{fct_generate_hex_colors}
\alias{fct_generate_hex_colors}
\title{Generate hex color codes for logFC values based on a given range.}
\usage{
fct_generate_hex_colors(df, range)
}
\arguments{
\item{df}{The data frame containing the 'logfc' values.}

\item{range}{A numeric vector of length 2 defining the color gradient scale.}
}
\value{
The input data frame with a new column, 'hex_from_scales'.
}
\description{
This function takes a data frame with a 'logfc' column and
generates a new column with hex color codes based on a specified range.
}
