% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fcPainter.R
\name{fct_get_chain_df}
\alias{fct_get_chain_df}
\title{Get a data frame of PDB chains from an API or local file.}
\usage{
fct_get_chain_df(pdb_code)
}
\arguments{
\item{pdb_code}{A string representing the PDB code (e.g., "7cid").}

\item{local_path}{The path to the directory where local PDB files are stored.}
}
\value{
A grouped `data.frame` with PDB chain information.
}
\description{
This function takes a PDB code, checks for a local JSON file,
and if not found, makes an API call to retrieve chain information.
}
