% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fcPainter.R
\name{fct_prepare_chainset}
\alias{fct_prepare_chainset}
\title{Combine and clean data to create a final color set for the viewer.}
\usage{
fct_prepare_chainset(joined_df)
}
\arguments{
\item{joined_df}{A dataframe from fct_join_logfc_chains (contains logfc, hex colors, chains).}
}
\value{
A `data.frame` with 'sele' and 'colorValue' columns for coloring the viewer.
}
\description{
This function merges the logFC data with the chain and dictionary information,
and formats it for use with NGLVieweR.
}
