% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fcPainter.R
\name{fct_prepare_legend_data}
\alias{fct_prepare_legend_data}
\title{Prepare data for the legend plot.}
\usage{
fct_prepare_legend_data(df, range)
}
\arguments{
\item{df}{The data frame containing the 'logfc' values.}

\item{range}{A numeric vector of length 2 defining the color gradient scale.}
}
\value{
A data frame with sorted logFC values and hex colors.
}
\description{
This function takes the uploaded logFC data and the color range,
and returns a data frame suitable for plotting a legend.
}
