% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fcPainter.R
\name{fct_read_logfc_file}
\alias{fct_read_logfc_file}
\title{Read a logFC CSV file uploaded by the user.}
\usage{
fct_read_logfc_file(file_path)
}
\arguments{
\item{file_path}{The temporary file path of the uploaded CSV.}
}
\value{
A `data.frame` of the uploaded data.
}
\description{
Reads a CSV file from a given path and returns it as a data frame.
}
