% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fcPainter.R
\docType{data}
\name{hg_data}
\alias{hg_data}
\title{Map gene symbols to UniProt IDs using AnnotationDbi.}
\format{
An object of class \code{list} of length 2.
}
\usage{
hg_data
}
\arguments{
\item{geneSymbol}{A character vector of gene names (e.g., 'SYMBOL').}
}
\value{
A `data.frame` with gene symbols and corresponding UniProt IDs.
}
\description{
This function takes a vector of gene names and returns a
data frame mapping them to UniProt IDs through HGNC IDs.
}
\keyword{datasets}
