% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fcPainter.R
\name{mod_fcPainter_ui}
\alias{mod_fcPainter_ui}
\title{PDBPainter UI Module}
\usage{
mod_fcPainter_ui(id)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}. DO NOT REMOVE.#'}
}
\description{
A UI module for the PDBPainter feature, including file upload
  and color range selection.
}
